+++ CHANGELOG.md
@@ -4050,3 +4050,4 @@
 - [2026-08-29 21:27] chore(test): update dependencies and cleanup unused imports in test
 - [2026-08-29 21:38] perf(analytics): optimize memory consumption in analytics list viewport
 - [2026-08-29 21:39] feat(network): integrate custom telemetry logger in network
+- [2026-08-29 21:52] refactor(analytics): optimize data flow between domain and data layers